LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_cross_origin_redirect_chain-expected.txt



PASS window.performance is defined
PASS window.performance.getEntriesByName is defined
PASS There should be one entry.
FAIL redirectStart > 0 in timing allowed cross-origin redirect. assert_true: redirectStart > 0 in timing allowed cross-origin redirect. expected true got false
FAIL redirectStart == startTime in timing allowed cross-origin redirect. assert_equals: redirectStart == startTime in timing allowed cross-origin redirect. expected <nondeterministic test output> but got 0
FAIL redirectEnd > redirectStart in timing allowed cross-origin redirect. assert_true: redirectEnd > redirectStart in timing allowed cross-origin redirect. expected true got false
PASS fetchStart >= redirectEnd in timing allowed cross-origin redirect.
